% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_meta_shell}
\alias{create_meta_shell}
\title{Create attribute shell}
\usage{
create_meta_shell(df, factor_cols = NULL, sep = ";")
}
\arguments{
\item{df}{a dataframe of data for which metadata attribute table is to be created}

\item{factor_cols}{character string or vector of names of columns containing factors}

\item{sep}{separator used to separate codes and definitions for each level in
the metadata table}
}
\value{
a dataframe consiting of one row per column of the input df. Columns
 represent minimum metadata requirements.
}
\description{
Create shell metadata table of variable attributes for a dataset. The function attempts to guess
guess the values where possible. The rest will need completing manually.
}
\examples{
\dontrun{
library(gapminder)
create_meta_shell(gapminder)
}
}
