% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_factors_meta}
\alias{get_factors_meta}
\title{Extract useful metadata from data and metadata tables}
\usage{
get_factors_meta(meta_tbl)
}
\arguments{
\item{meta_tbl}{a metadata table associated with df}
}
\value{
character vectors of variables names
}
\description{
Extract useful metadata from data and metadata tables
}
\details{
functions appended `df` extract information from a `df` of the data,
 while those appended `meta` from the `meta_tbl` containing the attribute metadata.
}
