% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_meta_shell}
\alias{get_meta_shell}
\title{Get attribute shell}
\usage{
get_meta_shell(df)
}
\arguments{
\item{df}{a dataframe of data for which metadata attribute table is to be created}
}
\value{
a dataframe consiting of one row per column of the input df. Columns
 represent minimum metadata requirements.
}
\description{
Create shell metadata table of variable attributes for a dataset. The function attempts to guess
guess the values where possible. The rest will need completing manually.
}
\examples{
\dontrun{
library(gapminder)
get_attr_shell(gapminder)
}
}
