% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_vars_df}
\alias{get_vars_df}
\title{Extract useful metadata from data and metadata tables}
\usage{
get_vars_df(df)
}
\arguments{
\item{df}{a dataframe of data for which metadata attribute table is to be created}
}
\value{
character vectors of variables names
}
\description{
Extract useful metadata from data and metadata tables
}
\details{
functions appended `df` extract information from a `df` of the data,
 while those appended `meta` from the `meta_tbl` containing the attribute metadata.
}
\examples{
\dontrun{
df <- gapminder::gapminder
get_vars_df(df)
get_vars_meta(meta_tbl)
get_factors_df(df)
get_factors_meta(meta_tbl)
}
}
