% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mt_create_meta_shell}
\alias{mt_create_meta_shell}
\title{Create metadata shell}
\usage{
mt_create_meta_shell(df, factor_cols = NULL, sep = ";")
}
\arguments{
\item{df}{a dataframe of data for which metadata attribute table is to be created}

\item{factor_cols}{character string or vector of names of columns containing factors
Code and level information is extracted auromatically from factor columns.}

\item{sep}{separator used to separate codes and definitions for each level in
the metadata table}
}
\value{
a dataframe consiting of one row per column of the input df. Columns
 represent minimum metadata requirements.
}
\description{
Create metadata table shell of variable attributes from a dataset. The function attempts to
guess values where possible. The rest require completing manually
}
\examples{
\dontrun{
library(gapminder)
create_meta_shell(gapminder)
}
}
