% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mt_extract_attr_tbl}
\alias{mt_extract_attr_tbl}
\title{Extract attribute table}
\usage{
mt_extract_attr_tbl(meta_tbl)
}
\arguments{
\item{meta_tbl}{a completed metadata table}
}
\value{
a dataframe containing attribute metadata in a format appropriate for passing to
 an EML attributeList
}
\description{
Extract attribute table to pass to EML attributeList
}
