% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz-utils.R
\name{mt_label}
\alias{mt_label}
\title{Get descriptive label}
\usage{
mt_label(meta_df, var)
}
\arguments{
\item{meta_df}{a metadatar metadata table}

\item{var}{character string. the variable for which to}
}
\value{
a character string descriptive label of the variable
}
\description{
Create a more descriptive label, including units where appropriate, for a
variable from information in the metadata table.
}
\examples{
\dontrun{
meta_df <- readr::read_csv(system.file("extdata", "gapminder_meta.csv", package="metadatar"))
mt_label(meta_df, var = "gdpPercap")
}
}
