% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mt_update_meta_tbl}
\alias{mt_update_meta_tbl}
\title{Update metadata table}
\usage{
mt_update_meta_tbl(df, meta_tbl, factor_cols = NULL, sep = ";")
}
\arguments{
\item{df}{a dataframe or tibble to which the metadata table relates to}

\item{meta_tbl}{a metadata table associated with df}

\item{factor_cols}{character string or vector of names of columns containing
factors (will only apply to new columns)}

\item{sep}{separator used to separate codes and definitions for each level in
the metadata table (must be the same used throughout meta_tbl)}
}
\value{
an updated meta_tbl with rows added for new df columns and any
deprecated rows trimmed.
}
\description{
Update the metadata table of a dataset. The function will identify variable columns in df
for which rows are missing in meta_tbl and append new rows for each. It will also
remove any deprecated rows which do not match columns in the dataset.
}
\concept{metadata table utilities}
